backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-JWT-Extended
python-dotenv
PyMySQL
cryptography
Werkzeug
gunicorn

Flask-Limiter
Flask-Talisman
Flask-Bcrypt
